Moved definitions of member static variables and functions into class
Moved definitions of member static variables into their classes by
declaring them to be inline, except for those that are impossible to do
so because they have the same type as their enclosing classes.

For consistensy, moved member function definitions into their classes,
except for those that are impossible to do so because of the order of
declarations.