Started moving *::generator::print functions
*::generator::print functions have been moved to their enclosing
classes of the inner classes they belong. For example, call
header::print instead of header::generator:print. Except for
event::generator::print, which have been moved to
event::print_registered_event_types. The difference is because there is
already an event::print function.

This change is to prepare for the removal of generator classes. An extra
derived_class_names member has been added to the affected classes. A
derived class should add its name to this list in its static
constructor(emulated by having an static instance of a dummy inner class
whose default constructor adds the enclosing name to the list).

Updated README.md to reflect the changes.