Fixed a problem where it's impossible to call IoT_ctrl_payload::increase
Since the getter of IoT_ctrl_packet for the payload returns a const
reference, it's impossible to call increase on the returned reference,
so IoT_ctrl_packet::increase_payload_counter has been added to allow
this.

Added a Usage Guides section to README.md to explain how to use the
improved code compared to the original.

Removed a obsolete line from README.md's Breaking Changes section.