Improved STATIC_CONSTRUCTOR to be usable in base classes
Improved STATIC_CONSTRUCTOR to be usable in base classes and multiple
times in the same class. However, it is unreliable. The static
constructor seems to only be called in non-template classes, including
non-template subclasses of template classes. Even then, the standard
doesn't even guarantee that it will be called in a non-template classes.